requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
fonttools==4.54.1
frozenlist==1.4.1
idna==3.10
kiwisolver==1.4.7
magic-filter==1.0.12
matplotlib==3.9.2
motor==3.6.0
multidict==6.1.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
yarl==1.14.0
